// src/enums/colors.ts
export enum COLORS {
    RED = "rgb(255,0,0)",
    WHITE = "rgb(255,255,255)",
    BLACK = "rgb(0,0,0)",
    NEWS = "rgb(255, 165, 0)",
    YELLOW = "rgba(255,255,0)",
    LIGHTBLUE = "rgba(51,204,255)",
    PURPLE = "rgb(128,0,128)",
    FADE1 = "rgba(230, 230, 230, 255)",
    FADE2 = "rgba(200, 200, 200, 0.78)",
    FADE3 =  "rgba(170, 170, 170, 0.67)",
    FADE4 = "rgba(140, 140, 140, 0.55)",
    FADE5 = "rgba(110, 110, 110, 0.43)",
    POWDER_BLUE = "rgba(182, 208, 226, 255)",
}